Add addon gathering logic
Takes all fingerprints of an addon directory and
fetches the files from Curse. If the addon exists,
it's an addon (duh), if not, it's a _custom_ addon. If
the _file_ exists (i.e. the fingerprint returns a value),
but the accompanying addon cannot be fetched in
bulk via /addon, then it's one that has been removed
from Curse but is still a bit more valid than a custom
addon.

Gathered addons are then compressed and stored locally
so that they don't need to be fetched on every app load.