Make worker work in packaged build
The nature of workers existing in an entirely different
process is a little spicy when it comes to the asar archive.
The easy answer would be to just disable asar in the build,
but this way we get to keep the (bulk of the) archiving, with
the worker still able to function.

All workers will need to do the same thing, so pulled into
a worker-utils file.