Correctly identify a matching curse addon file
Originally assumed the "gameVersionFlavor"
value was pointing to the game version that
a curse file was intended for. Works until
you get a file which is supported by multiple
versions.

Instead, this looks up in the sortableGameVersion
array which has a gameVersionTypeId which "appears"
to be fixed for retail wow, regardless of patch
version, to 517. Here's to hoping I don't need to
change this every time wow gets a major update.......